rename version to _version
