+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = shinylive
-version = attr: shinylive.version.SHINYLIVE_PACKAGE_VERSION
+version = attr: shinylive._version.SHINYLIVE_PACKAGE_VERSION
 author = Winston Chang
 author_email = winston@posit.co
 url = https://github.com/posit-dev/py-shinylive
